jar starts NetworkSimulator

+++ src/main/java/g7asmt1/server/Result.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class Result implements Serializable {
     private final String serviceName;
     private final int result;
-    private final int waitingTime;
-    private final int executionTime;
+    private final long waitingTime;
+    private final long executionTime;
     private final int zone;
 
-    public Result(String serviceName, int result, int waitingTime, int executionTime, int zone) {
+    public Result(String serviceName, int result, long waitingTime, long executionTime, int zone) {
         this.serviceName = serviceName;
         this.result = result;
         this.waitingTime = waitingTime;
